Controllers code restructuring

+++ internal/controllers/edge_controller.go
@@ -3,6 +3,7 @@ package controllers
 import (
 	"github.com/HarshKanjiya/escape-form-api/internal/services"
 	"github.com/HarshKanjiya/escape-form-api/internal/types"
+	"github.com/HarshKanjiya/escape-form-api/pkg/errors"
 	"github.com/HarshKanjiya/escape-form-api/pkg/utils"
 	"github.com/go-playground/validator/v10"
 	"github.com/gofiber/fiber/v2"
@@ -28,13 +29,19 @@ func NewEdgeController(service *services.EdgeService) *EdgeController {
 // @Success 200 {object} map[string]interface{}
 // @Router /edges [get]
 func (ec *EdgeController) Get(c *fiber.Ctx) error {
+
+	userId, ok := utils.GetUserId(c)
+	if ok == false {
+		return errors.Unauthorized("")
+	}
+
 	formId := c.Query("formId")
 	if formId == "" {
-		return utils.BadRequest(c, "formId is required")
+		return errors.BadRequest("formId is required")
 	}
 	edges, err := ec.edgeService.Get(c, formId)
 	if err != nil {
-		return utils.Error(c, fiber.StatusInternalServerError, "Failed to fetch edges")
+		return err
 	}
 	return utils.Success(c, edges, "Edges fetched successfully")
 }
@@ -47,12 +54,18 @@ func (ec *EdgeController) Get(c *fiber.Ctx) error {
 // @Success 200 {object} map[string]interface{}
 // @Router /edges [post]
 func (ec *EdgeController) Create(c *fiber.Ctx) error {
+
+	userId, ok := utils.GetUserId(c)
+	if ok == false {
+		return errors.Unauthorized("")
+	}
+
 	var edgeDto types.EdgeDto
 	if err := c.BodyParser(&edgeDto); err != nil {
-		return utils.BadRequest(c, "Invalid request body")
+		return errors.BadRequest("Invalid request body")
 	}
 	if err := ec.validator.Struct(&edgeDto); err != nil {
-		return utils.BadRequest(c, "Validation failed: "+err.Error())
+		return errors.BadRequest("Validation failed: " + err.Error())
 	}
 	edge, err := ec.edgeService.Create(c, &edgeDto)
 	if err != nil {
@@ -70,18 +83,24 @@ func (ec *EdgeController) Create(c *fiber.Ctx) error {
 // @Success 200 {object} map[string]interface{}
 // @Router /edges/{id} [patch]
 func (ec *EdgeController) Update(c *fiber.Ctx) error {
+
+	userId, ok := utils.GetUserId(c)
+	if ok == false {
+		return errors.Unauthorized("")
+	}
+
 	id := c.Params("id")
 	var edgeDto types.EdgeDto
 	if err := c.BodyParser(&edgeDto); err != nil {
-		return utils.BadRequest(c, "Invalid request body")
+		return errors.BadRequest("Invalid request body")
 	}
 	edgeDto.ID = id
 	if err := ec.validator.Struct(&edgeDto); err != nil {
-		return utils.BadRequest(c, "Validation failed: "+err.Error())
+		return errors.BadRequest("Validation failed: " + err.Error())
 	}
 	edge, err := ec.edgeService.Update(c, &edgeDto)
 	if err != nil {
-		return utils.Error(c, fiber.StatusInternalServerError, "Failed to update edge")
+		return err
 	}
 	return utils.Success(c, edge, "Edge updated successfully")
 }
@@ -95,6 +114,12 @@ func (ec *EdgeController) Update(c *fiber.Ctx) error {
 // @Success 200 {object} map[string]interface{}
 // @Router /edges/{id} [delete]
 func (ec *EdgeController) Delete(c *fiber.Ctx) error {
+
+	userId, ok := utils.GetUserId(c)
+	if ok == false {
+		return errors.Unauthorized("")
+	}
+
 	id := c.Params("id")
 	err := ec.edgeService.Delete(c, id)
 	if err != nil {
