+++ internal/controllers/project_controller.go
@@ -3,6 +3,7 @@ package controllers
 import (
 	"github.com/HarshKanjiya/escape-form-api/internal/services"
 	"github.com/HarshKanjiya/escape-form-api/internal/types"
+	"github.com/HarshKanjiya/escape-form-api/pkg/errors"
 	"github.com/HarshKanjiya/escape-form-api/pkg/utils"
 	"github.com/go-playground/validator/v10"
 	"github.com/gofiber/fiber/v2"
@@ -28,6 +29,12 @@ func NewProjectController(service services.IProjectService) *ProjectController {
 // @Success 200 {object} map[string]interface{}
 // @Router /projects [get]
 func (pc *ProjectController) Get(c *fiber.Ctx) error {
+
+	userId, ok := utils.GetUserId(c)
+	if ok == false {
+		return errors.Unauthorized("")
+	}
+
 	pagination := &types.PaginationQuery{
 		Page:   c.QueryInt("page", 1),
 		Limit:  c.QueryInt("limit", 10),
@@ -37,11 +44,13 @@ func (pc *ProjectController) Get(c *fiber.Ctx) error {
 	}
 
 	teamId := c.Query("teamId", "")
-	userId := c.Locals("user_id").(string)
+	if teamId == "" {
+		return errors.BadRequest("teamId is required")
+	}
 
 	projects, totalCount, err := pc.projectService.Get(c.Context(), userId, pagination, teamId)
 	if err != nil {
-		// 	return utils.InternalServerError(c, "Failed to fetch projects")
+		return err
 	}
 	return utils.Success(c, projects, "Projects fetched successfully", totalCount)
 }
@@ -55,11 +64,20 @@ func (pc *ProjectController) Get(c *fiber.Ctx) error {
 // @Success 200 {object} map[string]interface{}
 // @Router /projects/{projectId} [get]
 func (pc *ProjectController) GetById(c *fiber.Ctx) error {
-	projectId := c.Params("projectId")
-	userId := c.Locals("user_id").(string)
+
+	userId, ok := utils.GetUserId(c)
+	if ok == false {
+		return errors.Unauthorized("")
+	}
+
+	projectId := c.Params("projectId", "")
+	if projectId == "" {
+		return errors.BadRequest("projectId is required")
+	}
+
 	project, err := pc.projectService.GetById(c.Context(), userId, projectId)
 	if err != nil {
-		return utils.NotFound(c, "Project not found")
+		return errors.NotFound("Project")
 	}
 	return utils.Success(c, project, "Project fetched successfully")
 }
@@ -72,19 +90,24 @@ func (pc *ProjectController) GetById(c *fiber.Ctx) error {
 // @Success 200 {object} map[string]interface{}
 // @Router /projects [post]
 func (pc *ProjectController) Create(c *fiber.Ctx) error {
+
+	userId, ok := utils.GetUserId(c)
+	if ok == false {
+		return errors.Unauthorized("")
+	}
+
 	projectDto := new(types.ProjectDto)
 	if err := c.BodyParser(projectDto); err != nil {
-		return utils.BadRequest(c, "Invalid request body")
+		return errors.BadRequest("Invalid request body")
 	}
 
 	if err := pc.validator.Struct(projectDto); err != nil {
-		return utils.BadRequest(c, "Validation failed")
+		return errors.BadRequest("Validation failed")
 	}
 
-	userId := c.Locals("user_id").(string)
 	createdProject, err := pc.projectService.Create(c.Context(), userId, projectDto)
 	if err != nil {
-		return utils.BadRequest(c, "Failed to create project")
+		return err
 	}
 
 	return utils.Success(c, createdProject, "Project created successfully", 0)
@@ -99,16 +122,21 @@ func (pc *ProjectController) Create(c *fiber.Ctx) error {
 // @Success 200 {object} map[string]interface{}
 // @Router /projects/{id} [patch]
 func (pc *ProjectController) Update(c *fiber.Ctx) error {
+
+	userId, ok := utils.GetUserId(c)
+	if ok == false {
+		return errors.Unauthorized("")
+	}
+
 	projectDto := new(types.ProjectDto)
 	if err := c.BodyParser(projectDto); err != nil {
-		return utils.BadRequest(c, "Invalid request body")
+		return errors.BadRequest("Invalid request body")
 	}
 
 	if err := pc.validator.Struct(projectDto); err != nil {
-		return utils.BadRequest(c, "Validation failed")
+		return errors.BadRequest("Validation failed")
 	}
 
-	userId := c.Locals("user_id").(string)
 	ok, err := pc.projectService.Update(c.Context(), userId, &types.ProjectDto{
 		ID:          c.Params("id"),
 		Name:        projectDto.Name,
@@ -116,7 +144,7 @@ func (pc *ProjectController) Update(c *fiber.Ctx) error {
 		TeamID:      projectDto.TeamID,
 	})
 	if err != nil || !ok {
-		return utils.BadRequest(c, "Failed to update project")
+		return errors.BadRequest("Failed to update project")
 	}
 	return utils.Success(c, nil, "Project updated successfully")
 }
@@ -130,16 +158,21 @@ func (pc *ProjectController) Update(c *fiber.Ctx) error {
 // @Success 200 {object} map[string]interface{}
 // @Router /projects/{id} [delete]
 func (pc *ProjectController) Delete(c *fiber.Ctx) error {
+
+	userId, ok := utils.GetUserId(c)
+	if ok == false {
+		return errors.Unauthorized("")
+	}
+
 	projectId := c.Params("id")
 
 	if projectId == "" {
-		return utils.BadRequest(c, "Project ID is required")
+		return errors.BadRequest("Project ID is required")
 	}
 
-	userId := c.Locals("user_id").(string)
 	ok, err := pc.projectService.Delete(c.Context(), userId, projectId)
 	if err != nil || !ok {
-		return utils.BadRequest(c, "Failed to delete project")
+		return errors.BadRequest("Failed to delete project")
 	}
 	return utils.Success(c, nil, "Project deleted successfully")
 }
